checks if db exists

+++ database/migrations/2017_12_26_135333_create_posts_table.php
@@ -13,16 +13,18 @@ class CreatePostsTable extends Migration
      */
     public function up()
     {
-        Schema::create('posts', function (Blueprint $table) {
-            $table->increments('id')->index();
-            $table->string('name');
-            $table->string('url');
-            $table->integer('user_id')->unsigned();
-            $table->foreign('user_id')
+        if (!Schema::hasTable('password_resets')) {
+            Schema::create('posts', function (Blueprint $table) {
+                $table->increments('id')->index();
+                $table->string('name');
+                $table->string('url');
+                $table->integer('user_id')->unsigned();
+                $table->foreign('user_id')
                     ->references('id')->on('users')
                     ->onDelete('cascade');
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
